+++ src/Objects/AbstractConnection.php
@@ -6,20 +6,13 @@ use andrefelipe\Orchestrate\HttpClientInterface;
 /**
  *
  */
-abstract class AbstractConnection
+abstract class AbstractConnection implements ConnectionInterface
 {
     /**
      * @var HttpClientInterface
      */
     private $_httpClient;
 
-    /**
-     * Get current Http client instance.
-     *
-     * @param boolean $required
-     *
-     * @return HttpClientInterface
-     */
     public function getHttpClient($required = false)
     {
         if ($required) {
@@ -29,13 +22,6 @@ abstract class AbstractConnection
         return $this->_httpClient;
     }
 
-    /**
-     * Set the Http client which the object will use to make API requests.
-     *
-     * @param HttpClientInterface $httpClient
-     *
-     * @return HttpClientInterface self
-     */
     public function setHttpClient(HttpClientInterface $httpClient)
     {
         $this->_httpClient = $httpClient;